pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "aaqis"
version = "0.1.0"
description = "Air Quality Intelligence System for Astana"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Tim Seye"}
]

dependencies = [
    # Web Framework
    "django>=4.2",
    "djangorestframework>=3.14",
    "django-cors-headers>=4.3",
    
    # Database
    "psycopg2-binary>=2.9",
    
    # Data Processing
    "pandas>=2.0",
    "numpy>=1.24",
    
    # ML/DL
    "scikit-learn>=1.3",
    "tensorflow>=2.15",
    "statsmodels>=0.14",  # For STL decomposition
    
    # Visualization
    "plotly>=5.18",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    
    # API Clients
    "requests>=2.31",
    "httpx>=0.25",
    
    # Utilities
    "python-dotenv>=1.0",
    "celery>=5.3",
    "redis>=5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-django>=4.5",
    "pytest-cov>=4.1",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.7",
    "ipykernel>=6.27",
    "jupyter>=1.0",
]

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "src.core.settings"
python_files = ["test_*.py"]
